refactor: move spotify code to it's own file
